extract classification logic from component
